+++ bored.go
@@ -2,159 +2,41 @@ package main
 
 import (
 	"fmt"
-	"html"
-	"io/ioutil"
 	"log"
 	"strings"
 
 	"github.com/dambrisco/geddit"
-	"github.com/dambrisco/prose"
 	"github.com/jroimartin/gocui"
 	"github.com/toqueteos/webbrowser"
 
 	"github.com/dambrisco/bored/auth"
 	"github.com/dambrisco/bored/flag-parser"
-)
-
-type pagetype int
-
-const (
-	List pagetype = iota
-	Comments
-	Empty
-	Help
+	"github.com/dambrisco/bored/layout"
+	"github.com/dambrisco/bored/reddit"
 )
 
 var submissions []*geddit.Submission
-var views []*gocui.View
-var votes []*gocui.View
-var allViews []*gocui.View
+
 var session *geddit.LoginSession
 var opts geddit.ListingOptions
-var after string
 var count int
 var subreddit string
-var currentIndex int
-var currentPageType pagetype
-
-func load(limit int) []*geddit.Submission {
-	opts = geddit.ListingOptions{
-		Limit: limit,
-		After: after,
-	}
-	if subreddit == "" {
-		submissions, _ = session.Frontpage(geddit.DefaultPopularity, opts)
-	} else {
-		submissions, _ = session.SubredditSubmissions(subreddit, geddit.DefaultPopularity, opts)
-	}
-	return submissions
-}
-
-func layoutList(g *gocui.Gui) {
-	maxX, maxY := g.Size()
-	i := 0
-	for i < maxY-1 && i < len(submissions) {
-		name := fmt.Sprintf("submission-%d", i)
-		if v, err := g.SetView(fmt.Sprintf("vote-%d", i), -1, i, 1, i+2); err != nil && v != nil {
-			allViews = append(allViews, v)
-			votes = append(votes, v)
-			v.Frame = false
-			fmt.Fprint(v, "•")
-		}
-		if s, err := g.SetView(name, 1, i, maxX, i+2); err != nil && s != nil {
-			if i == currentIndex {
-				g.SetCurrentView(name)
-			}
-			views = append(views, s)
-			s.Frame = false
-			subm := submissions[i]
-			after = subm.Title
-			fmt.Fprint(s, buildTitleTag(subm))
-			allViews = append(allViews, s)
-		}
-		i += 1
-	}
-	setColor(g.CurrentView())
-}
-
-func layoutComments(g *gocui.Gui) {
-	s := submissions[currentIndex]
-	maxX, maxY := g.Size()
-	if vote, err := g.SetView("post-vote", -1, 0, 1, 2); err != nil && vote != nil {
-		allViews = append(allViews, vote)
-		vote.Frame = false
-		fmt.Fprint(vote, "•")
-	}
-	if title, err := g.SetView("post-title", 1, 0, maxX, 2); err != nil && title != nil {
-		allViews = append(allViews, title)
-		title.Frame = false
-		fmt.Fprint(title, buildTitleTag(s))
-	}
-	if meta, err := g.SetView("post-meta", -1, 2, maxX, 4); err != nil && meta != nil {
-		allViews = append(allViews, meta)
-		meta.Frame = false
-		fmt.Fprintf(meta, "Score: %d | Subreddit: %s", s.Score, s.Subreddit)
-	}
-	if text, err := g.SetView("post-text", -1, 4, maxX, maxY); err != nil && text != nil {
-		allViews = append(allViews, text)
-		text.Frame = false
-		text.Wrap = true
-		fmt.Fprintf(text, "%s", prose.Wrap(strings.Replace(html.UnescapeString(s.Selftext), "\n\n", "\n", -1), maxX))
-		g.SetCurrentView("post-text")
-	}
-}
-
-func layoutHelp(g *gocui.Gui) {
-	maxX, maxY := g.Size()
-	if help, err := g.SetView("help", -1, 0, maxX, maxY); err != nil && help != nil {
-		allViews = append(allViews, help)
-		help.Frame = false
-		help.Wrap = true
-		h, err := ioutil.ReadFile("help.txt")
-		l, err := ioutil.ReadFile("LICENSE")
-		if err != nil {
-			log.Panicln(err)
-		}
-		fmt.Fprintln(help, string(h))
-		fmt.Fprint(help, string(l))
-		g.SetCurrentView("help")
-	}
-}
-
-func layout(g *gocui.Gui) error {
-	maxX, count := g.Size()
-	count = count - 1
-	if title, err := g.SetView("title", -1, -1, maxX, 1); err != nil {
-		allViews = append(allViews, title)
-		title.Frame = false
-		title.BgColor = gocui.ColorBlue
-		title.FgColor = gocui.ColorWhite
-		fmt.Fprintln(title, "Bored v0.0.1")
-	}
-	if currentPageType == List {
-		layoutList(g)
-	} else if currentPageType == Comments {
-		layoutComments(g)
-	} else if currentPageType == Help {
-		layoutHelp(g)
-	}
-	return nil
-}
 
 func quit(g *gocui.Gui, _ *gocui.View) error {
 	return gocui.Quit
 }
 
 func help(g *gocui.Gui, v *gocui.View) error {
-	currentPageType = Help
-	clearViews(g, v)
+	layout.SetPage(layout.Help)
+	layout.Clear(g, v)
 	return nil
 }
 
 func cursorDown(g *gocui.Gui, v *gocui.View) error {
-	if currentPageType == List {
+	if layout.GetPage() == layout.List {
 		current := v.Name()
 		next := 0
+		views := layout.GetViews()
 		for i, v := range views {
 			if v.Name() == current {
 				next = i + 1
@@ -163,9 +45,9 @@ func cursorDown(g *gocui.Gui, v *gocui.View) error {
 		if next >= len(views) {
 			next = len(views) - 1
 		}
-		currentIndex = next
+		reddit.SetCurrentIndex(next)
 		return g.SetCurrentView(views[next].Name())
-	} else if currentPageType == Comments || currentPageType == Help {
+	} else if layout.GetPage() == layout.Comments || layout.GetPage() == layout.Help {
 		x, y := v.Origin()
 		v.SetOrigin(x, y+1)
 	}
@@ -173,9 +55,10 @@ func cursorDown(g *gocui.Gui, v *gocui.View) error {
 }
 
 func cursorUp(g *gocui.Gui, v *gocui.View) error {
-	if currentPageType == List {
+	if layout.GetPage() == layout.List {
 		current := v.Name()
 		prev := 0
+		views := layout.GetViews()
 		for i, v := range views {
 			if v.Name() == current {
 				prev = i - 1
@@ -184,9 +67,9 @@ func cursorUp(g *gocui.Gui, v *gocui.View) error {
 		if prev < 0 {
 			prev = 0
 		}
-		currentIndex = prev
+		reddit.SetCurrentIndex(prev)
 		return g.SetCurrentView(views[prev].Name())
-	} else if currentPageType == Comments || currentPageType == Help {
+	} else if layout.GetPage() == layout.Comments || layout.GetPage() == layout.Help {
 		x, y := v.Origin()
 		v.SetOrigin(x, y-1)
 	}
@@ -194,8 +77,10 @@ func cursorUp(g *gocui.Gui, v *gocui.View) error {
 }
 
 func upvote(g *gocui.Gui, v *gocui.View) error {
-	s := submissions[currentIndex]
-	if currentPageType == List {
+	s := reddit.GetCurrentSubmission()
+	currentIndex := reddit.GetCurrentIndex()
+	if layout.GetPage() == layout.List {
+		votes := layout.GetVotes()
 		vote := votes[currentIndex]
 		if vote.FgColor == gocui.ColorGreen {
 			vote.FgColor = gocui.ColorDefault
@@ -204,7 +89,7 @@ func upvote(g *gocui.Gui, v *gocui.View) error {
 			vote.FgColor = gocui.ColorGreen
 			session.Vote(s, geddit.UpVote)
 		}
-	} else if currentPageType == Comments {
+	} else if layout.GetPage() == layout.Comments {
 		vote, _ := g.View("post-vote")
 		if vote.FgColor == gocui.ColorGreen {
 			vote.FgColor = gocui.ColorDefault
@@ -218,8 +103,10 @@ func upvote(g *gocui.Gui, v *gocui.View) error {
 }
 
 func downvote(g *gocui.Gui, v *gocui.View) error {
-	s := submissions[currentIndex]
-	if currentPageType == List {
+	s := reddit.GetCurrentSubmission()
+	currentIndex := reddit.GetCurrentIndex()
+	if layout.GetPage() == layout.List {
+		votes := layout.GetVotes()
 		vote := votes[currentIndex]
 		if vote.FgColor == gocui.ColorRed {
 			vote.FgColor = gocui.ColorDefault
@@ -228,7 +115,7 @@ func downvote(g *gocui.Gui, v *gocui.View) error {
 			vote.FgColor = gocui.ColorRed
 			session.Vote(s, geddit.DownVote)
 		}
-	} else if currentPageType == Comments {
+	} else if layout.GetPage() == layout.Comments {
 		vote, _ := g.View("post-vote")
 		if vote.FgColor == gocui.ColorRed {
 			vote.FgColor = gocui.ColorDefault
@@ -242,62 +129,54 @@ func downvote(g *gocui.Gui, v *gocui.View) error {
 }
 
 func enter(g *gocui.Gui, v *gocui.View) error {
-	webbrowser.Open("https://www.reddit.com/" + submissions[currentIndex].Permalink)
+	submission := reddit.GetCurrentSubmission()
+	webbrowser.Open("https://www.reddit.com/" + submission.Permalink)
 	return nil
 }
 
 func link(g *gocui.Gui, v *gocui.View) error {
-	webbrowser.Open(submissions[currentIndex].URL)
+	submission := reddit.GetCurrentSubmission()
+	webbrowser.Open(submission.URL)
 	return nil
 }
 
 func front(g *gocui.Gui, v *gocui.View) error {
-	currentPageType = List
-	clearViews(g, v)
+	layout.SetPage(layout.List)
+	layout.Clear(g, v)
 	return nil
 }
 
 func comments(g *gocui.Gui, v *gocui.View) error {
-	currentPageType = Comments
-	clearViews(g, v)
+	layout.SetPage(layout.Comments)
+	layout.Clear(g, v)
 	return nil
 }
 
 func refresh(g *gocui.Gui, v *gocui.View) error {
-	if currentPageType == List {
-		currentPageType = Empty
-		currentIndex = 0
-		clearViews(g, v)
-		load(count)
+	if layout.GetPage() == layout.List {
+		layout.SetPage(layout.Empty)
+		reddit.SetCurrentIndex(0)
+		layout.Clear(g, v)
+		submissions = reddit.Load(count, subreddit, session)
 		return front(g, v)
-	} else if currentPageType == Comments || currentPageType == Help {
-		clearViews(g, v)
+	} else if layout.GetPage() == layout.Comments || layout.GetPage() == layout.Help {
+		layout.Clear(g, v)
 	}
 	return nil
 }
 
 func info(g *gocui.Gui, v *gocui.View) error {
-	if currentPageType == List {
-		s := submissions[currentIndex]
+	if layout.GetPage() == layout.List {
+		s := reddit.GetCurrentSubmission()
 		b := v.Buffer()
 		v.Clear()
 		if strings.HasPrefix(b, "S") {
-			fmt.Fprint(v, buildTitleTag(s))
+			fmt.Fprint(v, layout.BuildTitleTag(s))
 		} else {
 			fmt.Fprintf(v, "Score: %d | Subreddit: %s", s.Score, s.Subreddit)
 		}
 	}
 	return nil
-}
-
-func clearViews(g *gocui.Gui, v *gocui.View) {
-	for _, w := range allViews {
-		g.DeleteView(w.Name())
-	}
-	views = nil
-	votes = nil
-	allViews = nil
-	g.Flush()
 }
 
 func setKeybinds(g *gocui.Gui) {
@@ -345,34 +224,12 @@ func setKeybinds(g *gocui.Gui) {
 	}
 }
 
-func setColor(v *gocui.View) {
-	for _, w := range views {
-		w.FgColor = gocui.ColorDefault
-	}
-	v.FgColor = gocui.ColorBlue
-}
-
-func buildTitleTag(s *geddit.Submission) string {
-	title := html.UnescapeString(s.Title)
-	tag := "LINK"
-	if s.IsSelf {
-		tag = "SELF"
-	}
-	if s.IsNSFW {
-		tag = "NSFW+" + tag
-	}
-	if s.IsSticky {
-		tag = "STICKY+" + tag
-	}
-	return fmt.Sprintf("[%s] %s", tag, title)
-}
-
 func main() {
 	var err error
 	username, password, s := flagParser.Parse()
 	subreddit = s
 	session = auth.Login(username, password)
-	after = ""
+	reddit.SetAfter("")
 
 	g := gocui.NewGui()
 	if err := g.Init(); err != nil {
@@ -380,12 +237,12 @@ func main() {
 	}
 	_, count = g.Size()
 	count = count - 1
-	submissions = load(count)
-	currentPageType = List
+	submissions = reddit.Load(count, subreddit, session)
+	layout.SetPage(layout.List)
 	g.BgColor = gocui.ColorDefault
 	g.FgColor = gocui.ColorDefault
 	defer g.Close()
-	g.SetLayout(layout)
+	g.SetLayout(layout.Layout)
 	setKeybinds(g)
 
 	err = g.MainLoop()
